Update settings.json
I mistakenly commited this file because it was suggested in my VS Code.

But I've deleted the contents now. I can't really remember the default content of this file. Or should I delete the file permanently?

If you have an idea of what it's supposed to be like please let me know. 
Thanks!